feat: Add Hyprland theme configuration for Qt and GTK
This commit adds the necessary configuration files for the Hyprland theme in the `home-manager` module. Specifically, it includes the `qt.nix` and `gtk.nix` files, which set the theme and icon packages for Qt and GTK applications respectively. Additionally, the `catppuccin.nix` file has been updated to remove the GTK icon theme configuration, as it is now handled separately in the `gtk.nix` file.

The changes in this commit provide a consistent visual experience across Qt and GTK applications in the Hyprland theme.