feat: Update Hyprland theme configuration and dependencies
This commit updates the Hyprland theme configuration in the `configuration.nix` files for the `framework`, `wsl`, and `zephyrus` hosts. It includes the addition of various module imports and services related to the Hyprland theme, such as Bluetooth, OpenGL, Xbox, gamemode, gpg, steam, fprintd, openrgb, openssh, pipewire, and zram-generator.

The changes in this commit ensure that the Hyprland theme is properly configured and available across different hosts, providing a consistent visual experience for Qt and GTK applications.

Note: This commit builds upon the previous commits that added yt-dlp and the Hyprland theme configuration to the home-manager module.