feat: Update stylix configuration in catppuccin.nix
This commit updates the `stylix` configuration in the `catppuccin.nix` file. It enables the `nixvim.transparent_bg` target with the `main` and `sign_column` options set to `true`. Additionally, it sets the `polarity` to "dark" and updates the `opacity` settings for various applications and elements.

These changes enhance the visual appearance and functionality of the Catppuccin theme in the Home Manager module.